requirements.txt
psutil~=5.8.0
Pillow~=8.4.0
PyDirectInput~=1.0.4
PyAutoGUI~=0.9.53